Use more efficient way to assign MACs
With the old way, peak rate would drop to
~12Mpps when uddating either src or dst MAC.
With new way, peak rate is 14.88.

The new method is a copy of the MoonGen's lua
function (./include/protocol/ethernet.lua),
but this method works with octet ordering
as viewed by the user, not the reverse
order that MoonGen function uses.  This allows
us to increment the address natively.